requirements.txt
google-api-python-client==2.53.0
google-auth==2.6.6
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
python-dotenv==0.20.0
